reformat and zone cooldown
This is a slight reformat of function locations and a modification so each touched pad has an effect and the other pads fade slowly.  last touched pad stops having an effect.  ideally we would have the fade effect work as on the last pad as well.  for now this is good.